+++ tests/regression/redis/commands.py
@@ -794,23 +794,23 @@ class ServerCommandsTestCase(TestCase):
         self.make_zset('c', {'a1': 6, 'a4': 5, 'a5': 4})
 
         # sum, no weight
-        self.assert_(self.client.zunionstore('z', ['a', 'b', 'c']))
+        self.assertEqual(self.client.zunionstore('z', ['a', 'b', 'c']),5)
         self.assertEquals(
             list(self.client.zrange('z', 0, -1, withscores=True)),
             [(b'a2', 1), (b'a3', 3), (b'a5', 4), (b'a4', 7), (b'a1', 9)]
             )
 
         # max, no weight
-        self.assert_(
-            self.client.zunionstore('z', ['a', 'b', 'c'], aggregate='MAX')
-            )
+        self.assertEqual(
+            self.client.zunionstore('z', ['a', 'b', 'c'], aggregate='MAX'),5)
         self.assertEquals(
             list(self.client.zrange('z', 0, -1, withscores=True)),
             [(b'a2', 1), (b'a3', 2), (b'a5', 4), (b'a4', 5), (b'a1', 6)]
             )
 
         # with weight
-        self.assert_(self.client.zunionstore('z', {'a': 1, 'b': 2, 'c': 3}))
+        self.assertEqual(
+            self.client.zunionstore('z', {'a': 1, 'b': 2, 'c': 3}), 5)
         self.assertEquals(
             list(self.client.zrange('z', 0, -1, withscores=True)),
             [(b'a2', 1), (b'a3', 5), (b'a5', 12), (b'a4', 19), (b'a1', 23)]